Add files via upload
In this research, a hybrid method based on the combination of global and local information integration has been employed for identifying essential proteins in the PPI network. In the proposed method, to extract global information from the network, different sub-layers of the network with specific proteins are determined using the K-Shell decomposition method. Subsequently, to obtain local information in the network, the proposed ranking algorithm is utilized with the aim of identifying a set of influential proteins through information entropy.
In the method, initially, the information entropy and energy of each node (protein) are calculated as the initial spreading capability. Then, a node with the highest entropy and information energy is selected, and this process is iteratively repeated until a specific number of influential nodes (essential proteins) are chosen.